pyproject.toml
[project]
name = "databricks-mcp"
version = "0.2.0"
description = "A comprehensive MCP server for Databricks SQL, Jobs, and Metadata management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
authors = [
    {name = "Chris Cho", email = "alibuda90019@gmail.com"}
]
keywords = ["databricks", "mcp", "sql", "model-context-protocol", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
]
dependencies = [
    "databricks-sql-connector>=4.2.1",
    "databricks-sdk>=0.39.0",
    "fastmcp>=2.13.1",
    "python-dotenv>=1.2.1",
]

[project.urls]
Homepage = "https://github.com/ChrisChoTW/databricks-mcp"
Repository = "https://github.com/ChrisChoTW/databricks-mcp"
Issues = "https://github.com/ChrisChoTW/databricks-mcp/issues"
